+++ R/build-reference-index.R
@@ -48,7 +48,7 @@ data_reference_index_rows <- function(section, index, pkg) {
 
   if (has_name(section, "contents")) {
     check_all_characters(section$contents, index, pkg)
-    topics <- section_topics(section$contents, pkg)
+    topics <- section_topics(section$contents, pkg$topics, pkg$src_path)
 
     names <- topics$name
     topics$name <- NULL
@@ -141,9 +141,11 @@ check_missing_topics <- function(rows, pkg) {
   missing <- !in_index & !pkg$topics$internal
 
   if (any(missing)) {
+    topics <- paste0(pkg$topics$name[missing], collapse = ", ")
     abort(c(
       "All topics must be included in reference index",
-      paste0("Missing topics: ", paste0(pkg$topics$name[missing], collapse = ", "))
+      `x` = paste0("Missing topics: ", topics),
+      i = "Either add to _pkgdown.yml or use @keyword internal"
     ))
   }
 }